Fix: Ajout des besoins lors de la création de demande

+++ back/src/main/java/com/example/gestion_des_epi/gestion_epi/controller/PosteController.java
@@ -35,6 +35,6 @@ public class PosteController {
     }
     @DeleteMapping(path = "{id}")
     public String delete(@PathVariable int id) {
-        return this.posteService.delete(id);
+        return this.posteService.delete((long) id);
     }
 }
